Renamed bar-chart.js to histogram-chart.js
