+++ src/javascript/request-composer/components/histogram-chart.js
